+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 **unreleased**
 
 **unreleased**
-- [Feature] Remove the validation of the number of files in `BatchWriteDetectOperator`
+- [Feature] Create a new write detect operator with name `BatchWriteDetectSizeOnlyOperator` without the validation of the number of files
 - [Feature] Rename operator `BatchWriteProcessOperator` to `BatchWriteProcessNdjsonOperator`
 - [Feature] Create new operator `BatchWriteProcessOrcOperator` that write ORC file format
 - [Feature] Create new operator `BatchWriteProcessedMoveOperator`